Convert "UnitFactory" into an interface, create "DefaultUnitFactory"
